fix(product-baseline): extensionless-ELF manifest entries, linker scripts, scaffold cleanup, library-path validation, non-ELF additions
- pack_product_baseline: manifest classification now shares
  _discover_library_map's content-aware ELF predicate, so an extensionless
  ELF library gets a LibraryEntry instead of being silently omitted from
  the returned/persisted manifest.
- _discover_library_map: excludes GNU ld INPUT()/GROUP() linker scripts
  from discovery -- previously compared twice (directly and via its
  resolved target) or, when unresolvable, could abort the whole
  comparison.
- pack_product_baseline: wraps tempfile.mkstemp() in the same output-
  scaffold cleanup every other post-mkdir failure path already uses.
- unpack_product_baseline: validates each manifest library entry's
  declared path (containment + existence), the same way header roots
  already are, before publication.
- compare_product_directories: reports a non-ELF (.dll/.dylib) library
  addition as bundle_library_added, symmetric to the existing standalone-
  removal fallback -- compare_bundle()'s own addition detection is
  ELF-only.

Also splits tests/test_product_baseline.py's compare_product_directories
coverage into a new sibling file, test_product_baseline_compare.py, to
stay under the AI-readiness 2000-line hard cap.
